Add the function vc-p4-init-version, which always returns "1".
Add a fixed version of vc-register which uses vc-p4-init-version.

Add a fixed version of vc-mode-line which sets vc-mode to nil if a
file isn't under version control, so that the VC information is
removed from the mode line of a reverted file.

Fixv vc-p4-register so that it doesn't complain if the specified
revision is "1" or if the specified comment is an empty string.

Fix vc-p4-revert so that it does the right thing if the file being
reverted was added but not yet submitted (i.e., it isn't yet actually
in the repository).

Fix vc-p4-diff so that it does the right thing with a newly added file
(i.e., diffs the file against /dev/null).

[git-p4: depot-paths = "//guest/Ben_Key/vc-p4/,//guest/jonathan_kamens/vc-p4/,//guest/magnus_henoch/vc-p4/": change = 1253]