fix: vc-p4 doesn't recognize a file in its client because of case
When checking to see if a file is within a Perforce client, match the root
directory and current directory as reported by Perforce and ignore case
differences.

Sometimes (at least on Windows) the current directory and the client root can
point at the same directory, but because of the way Perforce creates directories
and files the case might differ.